Make Repository.stripWorkDir more robust
Repository.stripWorkDir was assuming too much about its File arguments,
namely that the given file was always a decendant of the workdir.
Futhermore, it did not "normalize" paths of relative files, but simpy
relied on the path returned by File.getPath().

The new behavior is to fall-back to using File.getAbsolutePath() if the
path returned by File.getPath() cannot be normalized. Test the new
behavior against mix of relative and absolute paths given as arguments.
Fixes problem with string out of bound exception when the path of the
given file is shorter than the workdir, usually meaning it is not a
decendant of the workdir.

Reported-by: Adam W. Hawks <awhawks@writeme.com>
Signed-off-by: Jonas Fonseca <fonseca@diku.dk>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>