Update license text to reference Eclipse Foundation, Inc.
The EDL specifically uses "Eclipse Foundation, Inc." as the legal
entity in bullet point 3 of the new-style BSD license.  We had in
the past used the generic "Git Development Community" for lack of
a better name here, but with the migration to the Eclipse umbrella
we need to be more precise.

This change has no impact on the terms applied to the code, and
thus does not require consent from the existing contributors.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>